+++ apps/studio/components/interfaces/Explorer/QueryResultTable.tsx
@@ -4,10 +4,10 @@ import { parseAsBoolean, useQueryState } from 'nuqs'
 import { Button, cn, Tooltip, TooltipContent, TooltipTrigger } from 'ui'
 
 import { subscriptionHasHipaaAddon } from '../Billing/Subscription/Subscription.utils'
-import { Results } from '../SQLEditor/UtilityPanel/Results'
 import { type QueryResult } from './types'
 import { AiAssistantDropdown } from '@/components/ui/AiAssistantDropdown'
 import CopyButton from '@/components/ui/CopyButton'
+import { DataGridResults } from '@/components/ui/DataGridResults'
 import { InlineLink, InlineLinkClassName } from '@/components/ui/InlineLink'
 import { useProjectSettingsV2Query } from '@/data/config/project-settings-v2-query'
 import { getSqlErrorLines } from '@/data/sql/utils'
@@ -186,5 +186,5 @@ const QueryError = ({
 
 // [Joshen] Eventually migrate the Results component here from SQL Editor
 const QueryResults = ({ rows }: { rows: NonNullable<QueryResult['rows']> }) => {
-  return <Results rows={rows} />
+  return <DataGridResults rows={rows} />
 }
